+++ tests/assets/config.php
@@ -1,5 +1,7 @@
 <?php
 return [
+    \Hawkbit\Application::KEY_ERROR => true,
+    \Hawkbit\Application::KEY_ERROR_CATCH => false,
     'templates' => [
         'default' => __DIR__ . '/../templates/default'
     ],