+++ tests/templates/acme/index.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: marco.bunge
+ * Date: 02.01.2017
+ * Time: 11:35
+ */
+
+// error response while testing
+if(!isset($this)){
+    return ;
+}
+
+echo $this->uppercase($text);